+++ e7awgsw/__init__.py
@@ -1,42 +1,5 @@
 __all__ = [
-    'AwgCtrl',
-    'CaptureCtrl',
-    'WaveSequence',
-    'CaptureParam',
-    'DspUnit', 
-    'CaptureUnit', 
-    'CaptureModule',
-    'DecisionFunc',
-    'CaptureParamElem',
-    'AWG',
-    'AwgErr',
-    'CaptureErr',
-    'AwgTimeoutError',
-    'CaptureUnitTimeoutError',
-    'AwgStartCmd',
-    'CaptureEndFenceCmd',
-    'WaveSequenceSetCmd',
-    'CaptureParamSetCmd',
-    'CaptureAddrSetCmd',
-    'FeedbackCalcOnClassificationCmd',
-    'WaveGenEndFenceCmd',
-    'AwgStartCmdErr',
-    'CaptureEndFenceCmdErr',
-    'WaveSequenceSetCmdErr',
-    'CaptureParamSetCmdErr',
-    'CaptureAddrSetCmdErr',
-    'FeedbackCalcOnClassificationCmdErr',
-    'WaveGenEndFenceCmdErr',
-    'SequencerCtrl',
-    'dsp']
+    'feedback',
+    'simplemulti',
+]
 
-from e7awgsw.hwdefs import DspUnit, CaptureUnit, CaptureModule, DecisionFunc, CaptureParamElem, AWG, FeedbackChannel, AwgErr, CaptureErr, SequencerErr
-from e7awgsw.awgctrl import AwgCtrl
-from e7awgsw.capturectrl import CaptureCtrl
-from e7awgsw.wavesequence import WaveSequence
-from e7awgsw.captureparam import CaptureParam
-from e7awgsw.sequencercmd import AwgStartCmd, CaptureEndFenceCmd, WaveSequenceSetCmd, CaptureParamSetCmd, CaptureAddrSetCmd, FeedbackCalcOnClassificationCmd, WaveGenEndFenceCmd
-from e7awgsw.sequencercmd import AwgStartCmdErr, CaptureEndFenceCmdErr, WaveSequenceSetCmdErr, CaptureParamSetCmdErr, CaptureAddrSetCmdErr, FeedbackCalcOnClassificationCmdErr, WaveGenEndFenceCmdErr
-from e7awgsw.sequencerctrl import SequencerCtrl
-from e7awgsw.exception import AwgTimeoutError, CaptureUnitTimeoutError
-from e7awgsw.dspmodule import dsp
