+++ emulator/feedback/awg.py
@@ -2,9 +2,9 @@ import threading
 import struct
 from enum import IntEnum
 from e7awgsw import WaveSequence
-from e7awgsw.memorymap import WaveParamRegs
-from e7awgsw.hwparam import WAVE_SAMPLE_SIZE
-from e7awgsw.logger import get_file_logger, get_stderr_logger, log_error
+from e7awgsw import WaveParamRegs
+from e7awgsw import WAVE_SAMPLE_SIZE
+from e7awgsw import get_file_logger, get_stderr_logger, log_error
 
 
 class Awg:
