+++ emulator/feedback/awgcontroller.py
@@ -1,8 +1,8 @@
 import awg
 from register import RwRegister, RoRegister
 from e7awgsw import AWG
-from e7awgsw.memorymap import AwgMasterCtrlRegs, AwgCtrlRegs, WaveParamRegs
-from e7awgsw.logger import get_file_logger, get_stderr_logger, log_error
+from e7awgsw import AwgMasterCtrlRegs, AwgCtrlRegs, WaveParamRegs
+from e7awgsw import get_file_logger, get_stderr_logger, log_error
 
 
 class AwgController(object):
