+++ sphinx_build.sh
@@ -4,5 +4,5 @@ mkdir ./docs/_build
 mv ./docs/index.rst .
 rm -f ./docs/*.rst
 mv ./index.rst ./docs/index.rst
-sphinx-apidoc --module-first -f  --implicit-namespaces -o ./docs ./e7awgsw/ ./e7awgsw/__init__.py ./e7awgsw/memorymap.py ./e7awgsw/udpaccess.py ./e7awgsw/uplpacket.py ./e7awgsw/hwparam.py ./e7awgsw/logger.py ./e7awgsw/classification.py ./e7awgsw/lock.py ./e7awgsw/dspmodule.py
+sphinx-apidoc --module-first -f  --implicit-namespaces -o ./docs ./e7awgsw/ ./e7awgsw/__init__.py ./e7awgsw/feedback/memorymap.py ./e7awgsw/feedback/udpaccess.py ./e7awgsw/feedback/uplpacket.py ./e7awgsw/feedback/hwparam.py ./e7awgsw/feedback/logger.py ./e7awgsw/feedback/classification.py ./e7awgsw/feedback/lock.py ./e7awgsw/feedback/dspmodule.py
 sphinx-build ./docs ./docs/_build
